Refactor database and repository into data module, fix circular dependency
