+++ app/build.gradle
@@ -118,6 +118,7 @@ dependencies {
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.2.0'
 
     implementation(project(":core:ui"))
-    implementation(project(":core:translations")) // RECHECK
+    implementation(project(":core:translations"))
     implementation(project(":core:util"))
+    implementation(project(":data:mongo"))
 }
