+++ core/ui/build.gradle.kts
@@ -42,13 +42,10 @@ dependencies {
     implementation(libs.material3.compose)
     implementation(libs.compose.tooling.preview)
 
-    implementation(libs.coil)
-
     implementation(libs.realm.sync)
     implementation(libs.coroutines.core)
 
     implementation(libs.desugar.jdk)
 
     implementation(project(":core:translations"))
-    implementation(project(":core:util"))
 }
