+++ core/util/src/main/java/com/xdiach/util/UtilFunctions.kt
@@ -3,11 +3,7 @@
 package com.xdiach.util
 
 import android.net.Uri
-import androidx.core.net.toUri
 import com.google.firebase.storage.FirebaseStorage
-import com.google.firebase.storage.ktx.storageMetadata
-import com.xdiach.diarymoodapp.data.database.entity.ImageToDeleteEntity
-import com.xdiach.diarymoodapp.data.database.entity.ImageToUploadEntity
 import io.realm.kotlin.types.RealmInstant
 import java.time.Instant
 
@@ -32,27 +28,6 @@ fun fetchImagesFromFirebase(
             }
         }
     }
-}
-
-fun retryUploadingImageToFirebase(
-    imageToUpload: ImageToUploadEntity,
-    onSuccess: () -> Unit
-) {
-    val storage = FirebaseStorage.getInstance().reference
-    storage.child(imageToUpload.remoteImagePath).putFile(
-        imageToUpload.imageUri.toUri(),
-        storageMetadata { },
-        imageToUpload.sessionUri.toUri()
-    ).addOnSuccessListener { onSuccess() }
-}
-
-fun retryDeletingImageFromFirebase(
-    imageToDelete: ImageToDeleteEntity,
-    onSuccess: () -> Unit
-) {
-    val storage = FirebaseStorage.getInstance().reference
-    storage.child(imageToDelete.remoteImagePath).delete()
-        .addOnSuccessListener { onSuccess() }
 }
 
 fun RealmInstant.toInstant(): Instant {
