+++ data/mongo/build.gradle.kts
@@ -26,6 +26,7 @@ android {
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
+        isCoreLibraryDesugaringEnabled = true
     }
     kotlinOptions {
         jvmTarget = "1.8"
@@ -44,4 +45,8 @@ dependencies {
 
     implementation(libs.hilt.android)
     implementation(libs.hilt.compiler)
+
+    implementation(libs.desugar.jdk)
+
+    implementation(project(":core:util"))
 }
