+++ data/mongo/src/main/java/com/xdiach/mongo/database/entity/ImageToUploadEntity.kt
@@ -1,8 +1,8 @@
-package com.xdiach.diarymoodapp.data.database.entity
+package com.xdiach.mongo.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.xdiach.diarymoodapp.util.Constants.IMAGE_TO_UPLOAD_TABLE
+import com.xdiach.util.Constants.IMAGE_TO_UPLOAD_TABLE
 
 @Entity(tableName = IMAGE_TO_UPLOAD_TABLE)
 data class ImageToUploadEntity(
